Moved package home page from https://github.com/iNamik/container.go to https://github.com/iNamik/go_container
It turns out that Go's clean functionaility can be dangerious if you have a non-derived file with same exact name as the containing package folder.  This is not currently a problem with the container project, but having a project folder name ending in '.go' runs the risk of 'cleaning' src files with the same name.  So I am moving all of my Go projects from '*.go' to 'go_*'